requirements.txt
openai==0.27.2
langchain==0.0.208
chromadb==0.4.3
tk==0.1.0
python-dotenv==1.0.0
tiktoken